Fix C warning. Minor updates to ohmic heating integral
I took a look at how PLUTO was computing the volume elements and updated the code to use its numerical method since it is faster than the naive one I wrote previously. The actual values are not significantly different (the same for up to about 4 non-zero digits).

The volume elements in the Python code are quite comparable to those in PLUTO. For example, (i,j,k)=(10,37,41) has a volume element of 3.287376861480138e-05 in Python and 0.000030 in PLUTO. Although this is slightly different, it is the same order of magnitude and I'm not too concerned about it. However, the ohmic heating computed in PLUTO is still higher than that computed in Python, by about 1.5 times in this run. I don't think the difference in volume element values is significant enough to explain this difference, so this still needs to be debugged further. I also tried removing the averaging from the field quantities (J, eta), but the plot looks the same. The integration log from PLUTO is included in this commit